Update TryParseMethodCache and fix build error - WIP
